requirements-light.txt
# Core dependencies - lightweight version
flask==2.3.3
flask-socketio==5.3.4
flask-login==0.6.2
flask-wtf==1.1.1
flask-sqlalchemy==3.0.5
flask-migrate==4.0.4
python-dotenv==1.0.0

# Simplified NLP Engine - using smaller models
nltk==3.8.1
scikit-learn==1.3.0

# Lightweight File Management
werkzeug==3.1.3

# Security
bcrypt==4.0.1
PyJWT==2.8.0

# Testing and Logging
pytest==7.4.0
pytest-flask==1.2.0

# Utilities
markdown==3.4.4
requests==2.31.0
